+++ src/main/java/com/sandymandy/pleasurecraft/PleasureCraft.java
@@ -3,7 +3,7 @@ package com.sandymandy.pleasurecraft;
 import com.sandymandy.pleasurecraft.entity.PleasureCraftEntities;
 import com.sandymandy.pleasurecraft.item.PleasureCraftItemGroups;
 import com.sandymandy.pleasurecraft.item.PleasureCraftItems;
-import com.sandymandy.pleasurecraft.network.PleasureCraftPackets;
+import com.sandymandy.pleasurecraft.networking.PleasureCraftPackets;
 import com.sandymandy.pleasurecraft.screen.GirlInventoryScreenHandler;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.screenhandler.v1.ExtendedScreenHandlerType;
@@ -20,10 +20,17 @@ public class PleasureCraft implements ModInitializer {
 	public static final String MOD_ID = "pleasurecraft";
 	public static final Logger LOGGER = LoggerFactory.getLogger("PleasureCraft");
 
-	public static final ExtendedScreenHandlerType<GirlInventoryScreenHandler, GirlScreenData> GIRL_SCREEN_HANDLER =
+	public static final ExtendedScreenHandlerType<GirlInventoryScreenHandler, GirlScreenData> GIRL_INVENTORY_SCREEN_HANDLER =
 			Registry.register(
 					Registries.SCREEN_HANDLER,
-					Identifier.of(PleasureCraft.MOD_ID, "lucy_screen"),
+					Identifier.of(PleasureCraft.MOD_ID, "girl_inventory_screen"),
+					new ExtendedScreenHandlerType<>(GirlInventoryScreenHandler::new, GirlScreenData.PACKET_CODEC)
+			);
+
+	public static final ExtendedScreenHandlerType<GirlInventoryScreenHandler, GirlScreenData> GIRL_TALK_SCREEN_HANDLER =
+			Registry.register(
+					Registries.SCREEN_HANDLER,
+					Identifier.of(PleasureCraft.MOD_ID, "girl_talk_screen"),
 					new ExtendedScreenHandlerType<>(GirlInventoryScreenHandler::new, GirlScreenData.PACKET_CODEC)
 			);