+++ src/main/java/com/sandymandy/pleasurecraft/client/PleasureCraftKeybinds.java
@@ -1,10 +1,6 @@
 package com.sandymandy.pleasurecraft.client;
 
-import com.sandymandy.pleasurecraft.network.players.CumKeybindC2SPacket;
-import com.sandymandy.pleasurecraft.network.players.ThrustKeybindC2SPacket;
-import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
-import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.minecraft.client.option.KeyBinding;
 import net.minecraft.client.util.InputUtil;
 import org.lwjgl.glfw.GLFW;