+++ src/main/java/com/sandymandy/pleasurecraft/entity/girls/LucyEntity.java
@@ -1,5 +1,6 @@
 package com.sandymandy.pleasurecraft.entity.girls;
 
+import com.sandymandy.pleasurecraft.scene.SceneOption;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.attribute.DefaultAttributeContainer;
@@ -7,6 +8,7 @@ import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.mob.MobEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.Items;
+import net.minecraft.text.Text;
 import net.minecraft.world.World;
 
 import java.util.List;
@@ -38,6 +40,16 @@ public class LucyEntity extends AbstractGirlEntity{
 
     @Override
     public float getYAxisGUI(){return 0.0525F;}
+
+    @Override
+    public List<SceneOption> getSceneOptions() {
+        return List.of(
+                new SceneOption("Paizuri",
+                        "paizuri_start", List.of("paizuri_slow"),List.of("paizuri_fast"),"paizuri_cum"),
+                new SceneOption("Blow Job",
+                        "blowjob_start",List.of("blowjob_slow"),List.of("blowjob_fast"),"blowjob_cum")
+        );
+    }
 
     @Override
     protected Map<EquipmentSlot, List<String>> getClothingBones() {