+++ src/main/java/com/sandymandy/pleasurecraft/networking/C2S/AnimationSyncC2SPacket.java
@@ -1,4 +1,4 @@
-package com.sandymandy.pleasurecraft.network.girls;
+package com.sandymandy.pleasurecraft.networking.C2S;
 
 import com.sandymandy.pleasurecraft.PleasureCraft;
 import net.minecraft.network.RegistryByteBuf;