+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.4
+- Added a custom GUI that gives you the sex options that are girl depended.
+- Make the entity stop and look at the player when interacting
+
 ## 0.3.1
 - Added a change log to keep track of changes each update
 - Made it so that the armor applied to the entity effects how many hits it can take.
@@ -11,3 +15,8 @@
 - Made it so that the Scene System Runs on the Server. Before it started on the server then switched to the client after finishing the intro.
 - Made it so that you can finish the Scene after starting it after reaching a serten amout of progress.
 - Added the ability to thrust in the Scene.
+
+## TODO
+- Fix the animation randomizer in the Scene
+- Make it so that the player needs a bed to start bed related scenes. By using this.getMoveControl().moveTo(x, y, z, s);
+- Make the AI more advance.